Ensure we execute grml-autoconfig
To run sheep at startup we are supposed to use the kernel paramater
services=sheep. If it does not work for any reason we can use
startup=/usr/bin/sheep-service instead.

As a side effect, the script grml-autoconfig is not executed anymore,
the sheep-service script is executed instead. This is why we run
grml-autoconfig from sheep-service. It ensures, everything is well
configured, like ssh service and user password.